third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/html-or-svg-or-mathml-element-interfaces-expected.txt

This is a testharness.js-based test.
[FAIL] MathML a element should have a tabIndex value of 0
  assert_equals: expected 0 but got -1
Harness: the test ran to completion.
